Extracted checkValidity as separate utility function
