+++ src/containers/Auth/Auth.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { Redirect } from "react-router-dom";
+import { checkValidity } from "../../shared/utility";
 
 import Input from "../../components/UI/Input/Input";
 import Button from "../../components/UI/Button/Button";
@@ -56,7 +57,7 @@ class Auth extends Component {
             [controlName]: {
                 ...this.state.controls[controlName],
                 value: event.target.value,
-                valid: !this.checkValidity(
+                valid: !checkValidity(
                     event.target.value,
                     this.state.controls[controlName].validation
                 ),
@@ -67,26 +68,7 @@ class Auth extends Component {
         this.setState({ controls: updatedControls });
     };
 
-    checkValidity(value, rules) {
-        let isValid = true;
-        if (!rules) {
-            return true;
-        }
-        if (rules.required) {
-            isValid = value.trim() !== "" && isValid;
-        }
-        if (rules.maxLength) {
-            isValid = value.length <= rules.maxLength && isValid;
-        }
-        if (rules.minLength) {
-            isValid = value.length >= rules.minLength && isValid;
-        }
-        if (rules.isEmail) {
-            const atPresent = value.indexOf("@") > 0;
-            isValid = atPresent && isValid;
-        }
-        return isValid;
-    }
+    
 
     submitHandler = (event) => {
         event.preventDefault();
