+++ src/containers/Checkout/ContactData/ContactData.js
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import axios from "../../../axios-orders";
 import { connect } from 'react-redux';
 import * as actions from "../../../store/acitons/index";
+import { checkValidity } from "../../../shared/utility";
 
 import Button from "../../../components/UI/Button/Button";
 import Spinner from "../../../components/UI/Spinner/Spinner";
@@ -120,7 +121,7 @@ class ContactData extends Component {
 
         updatedFormElement.value = event.target.value;
         updatedOrderForm[inputId] = updatedFormElement;
-        updatedFormElement.valid = this.checkValidity(updatedFormElement.value, updatedFormElement.validation);
+        updatedFormElement.valid = checkValidity(updatedFormElement.value, updatedFormElement.validation);
         updatedFormElement.touched = true;
 
         let formIsValid = true;
@@ -131,23 +132,6 @@ class ContactData extends Component {
 
         this.setState({ orderForm: updatedOrderForm, formIsValid: formIsValid });
     };
-
-    checkValidity(value, rules) {
-        let isValid = true;
-        if(!rules) {
-            return true;
-        }
-        if(rules.required) {
-            isValid = value.trim() !== '' && isValid;
-        }
-        if(rules.maxLength) {
-            isValid = value.length >= rules.maxLength && isValid;
-        }
-        if(rules.minLength) {
-            isValid = value.length <= rules.minLength && isValid;
-        }
-        return isValid;
-    }
 
     render() {
         const formElementArray = [];
