+++ src/shared/utility.js
@@ -1,0 +1,20 @@
+export const checkValidity = (value, rules) => {
+    let isValid = true;
+    if (!rules) {
+        return true;
+    }
+    if (rules.required) {
+        isValid = value.trim() !== "" && isValid;
+    }
+    if (rules.maxLength) {
+        isValid = value.length <= rules.maxLength && isValid;
+    }
+    if (rules.minLength) {
+        isValid = value.length >= rules.minLength && isValid;
+    }
+    if (rules.isEmail) {
+        const atPresent = value.indexOf("@") > 0;
+        isValid = atPresent && isValid;
+    }
+    return isValid;
+}
